# perform arrayToQPath in chunks