This PR optimizes ```arrayToQPath()``` with the following techniques:
1) creation of path in chunks
   * this reduces the amount of temporary memory being occupied by QPolygonF
2) reserving memory used by QPainterPath.
   * internally, elements are appended into QPainterPath one-by-one.
   * reserving avoids memory re-allocation and the resultant copying
3) deferring patching of non-finite values
   * this avoids creation of new copies of input data x and y by patching directly into the final buffer

The code was restructured and refactored to allow easier reasoning and to prevent changes in one mode from affecting another mode. Despite the larger code, I believe that the code is easier to follow now than before.

### master (PySide6 / Windows)
```
[ 75.00%] ··· arrayToQPath.TimeSuiteAllFinite.time_test
[ 75.00%] ··· ========= ============= ============ ============= ============
              --                                param2
              --------- -----------------------------------------------------
                param1       all         finite        pairs        array
              ========= ============= ============ ============= ============
                10000     90.7±0.3μs   95.7±0.5μs   1.03±0.01ms    1.06±0ms
                100000   2.35±0.02ms   2.50±0.1ms   11.9±0.09ms   12.0±0.1ms
               1000000    26.0±0.3ms   26.8±0.1ms    121±0.8ms    119±0.7ms
              ========= ============= ============ ============= ============

[100.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                                          ok
[100.00%] ··· ========= ============= ============= ============ =============
              --                                param2
              --------- ------------------------------------------------------
                param1       all          finite       pairs         array
              ========= ============= ============= ============ =============
                10000      448±2μs       181±1μs      1.52±0ms      1.54±0ms
                100000   6.49±0.02ms   3.06±0.01ms   17.5±0.1ms   17.6±0.08ms
               1000000    65.2±0.3ms    33.1±0.2ms   175±0.5ms     174±0.3ms
              ========= ============= ============= ============ =============
```

### this PR (PySide6 / Windows)
```
[ 75.00%] ··· arrayToQPath.TimeSuiteAllFinite.time_test                                                              ok
[ 75.00%] ··· ========= ============= ============= ============= =============
              --                                 param2
              --------- -------------------------------------------------------
                param1       all          finite        pairs         array
              ========= ============= ============= ============= =============
                10000     81.4±0.2μs    91.1±0.4μs   1.06±0.01ms   1.10±0.03ms
                100000   1.31±0.02ms   1.41±0.03ms   10.9±0.04ms   11.1±0.04ms
               1000000   13.4±0.06ms   14.1±0.07ms    112±0.6ms     110±0.3ms
              ========= ============= ============= ============= =============

[100.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                                          ok
[100.00%] ··· ========= ============= ============= ============= =============
              --                                 param2
              --------- -------------------------------------------------------
                param1       all          finite        pairs         array
              ========= ============= ============= ============= =============
                10000      179±5μs      115±0.2μs    1.16±0.03ms   1.18±0.01ms
                100000   2.59±0.03ms   1.37±0.01ms    12.2±0.1ms   12.3±0.09ms
               1000000    21.9±0.1ms    14.6±0.2ms    125±0.4ms     122±0.2ms
              ========= ============= ============= ============= =============
```