@j9ac9k 2021-08-12
@pijyoi the diff looks good, on macOS I'm not seeing any significant difference in performance, which is somewhat odd.  I'll test on my windows machine as well.

For the sake of my record, after cherry-picking the benchmark file, and running on master, these are the results I got:



master

```
[ 37.50%] ··· arrayToQPath.TimeSuiteAllFinite.time_test
[ 37.50%] ··· ========= ============= ============= ============ ============
              --                                param2
              --------- -----------------------------------------------------
                param1       all          finite       pairs        array
              ========= ============= ============= ============ ============
                10000      98.4±5μs      110±5μs      917±50μs     865±7μs
                100000   1.18±0.03ms   1.26±0.03ms   9.61±0.6ms   9.25±0.5ms
               1000000     23.4±2ms     23.2±0.3ms    103±3ms      103±2ms
              ========= ============= ============= ============ ============

[ 50.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                                                                                                          ok
[ 50.00%] ··· ========= ============ ============= ============= =============
              --                                param2
              --------- ------------------------------------------------------
                param1      all          finite        pairs         array
              ========= ============ ============= ============= =============
                10000     528±20μs      233±30μs    1.47±0.07ms   1.46±0.08ms
                100000   4.69±0.1ms   1.56±0.04ms    13.6±0.3ms    14.0±0.2ms
               1000000    61.7±2ms     26.6±0.9ms     154±6ms       150±2ms
              ========= ============ ============= ============= =============

[ 50.00%] ·· Building for virtualenv-py3.8-numpy-pyside2
[ 50.00%] ·· Benchmarking virtualenv-py3.8-numpy-pyside2
[ 62.50%] ··· Running (arrayToQPath.TimeSuiteAllFinite.time_test--)..
[ 87.50%] ··· arrayToQPath.TimeSuiteAllFinite.time_test                                                                                                                              ok
[ 87.50%] ··· ========= ============= ============= ============ ============
              --                                param2
              --------- -----------------------------------------------------
                param1       all          finite       pairs        array
              ========= ============= ============= ============ ============
                10000      106±6μs       104±2μs      942±40μs     847±5μs
                100000   1.16±0.02ms   1.25±0.02ms   9.59±0.2ms   9.45±0.5ms
               1000000    22.0±0.3ms    22.7±0.3ms    101±2ms      100±3ms
              ========= ============= ============= ============ ============

[100.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                                                                                                          ok
[100.00%] ··· ========= ============= ============ ============= ============
              --                                param2
              --------- -----------------------------------------------------
                param1       all         finite        pairs        array
              ========= ============= ============ ============= ============
                10000      456±20μs     159±3μs     1.37±0.02ms   1.36±0.1ms
                100000   4.54±0.05ms   1.72±0.2ms    15.2±0.6ms    14.6±1ms
               1000000     61.7±2ms    24.9±0.7ms     150±2ms      151±3ms
              ========= ============= ============ ============= ============
```

on qpath_chunks

```
[ 37.50%] ··· arrayToQPath.TimeSuiteAllFinite.time_test                                                                                                                              ok
[ 37.50%] ··· ========= ============ ============= ========== ============
              --                              param2
              --------- --------------------------------------------------
                param1      all          finite      pairs       array
              ========= ============ ============= ========== ============
                10000     113±20μs      123±20μs    948±40μs    995±40μs
                100000   1.31±0.2ms   1.43±0.07ms   10.0±1ms   10.2±0.4ms
               1000000   25.5±0.6ms     26.8±1ms    107±2ms     106±4ms
              ========= ============ ============= ========== ============

[ 50.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                                                                                                          ok
[ 50.00%] ··· ========= ============ ============ ============= ============
              --                               param2
              --------- ----------------------------------------------------
                param1      all         finite        pairs        array
              ========= ============ ============ ============= ============
                10000     455±10μs     170±10μs    1.44±0.06ms   1.49±0.2ms
                100000   4.87±0.1ms   1.95±0.7ms    14.4±0.4ms    13.7±1ms
               1000000   64.1±0.7ms    28.1±1ms      155±3ms      151±2ms
              ========= ============ ============ ============= ============

[ 50.00%] ·· Building for virtualenv-py3.8-numpy-pyside2
[ 50.00%] ·· Benchmarking virtualenv-py3.8-numpy-pyside2
[ 62.50%] ··· Running (arrayToQPath.TimeSuiteAllFinite.time_test--)..
[ 87.50%] ··· arrayToQPath.TimeSuiteAllFinite.time_test                                                                                                                              ok
[ 87.50%] ··· ========= ============= ============= ============ ============
              --                                param2
              --------- -----------------------------------------------------
                param1       all          finite       pairs        array
              ========= ============= ============= ============ ============
                10000      92.8±3μs      107±4μs      896±10μs     863±20μs
                100000   1.23±0.05ms   1.27±0.01ms   9.38±0.3ms   9.24±0.6ms
               1000000    22.5±0.9ms    23.3±0.6ms    102±2ms      99.1±1ms
              ========= ============= ============= ============ ============

[100.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                                                                                                          ok
[100.00%] ··· ========= ============ ============= ============= =============
              --                                param2
              --------- ------------------------------------------------------
                param1      all          finite        pairs         array
              ========= ============ ============= ============= =============
                10000     439±9μs       156±3μs     1.36±0.03ms   1.43±0.08ms
                100000   4.65±0.1ms   1.51±0.05ms    13.6±0.3ms    13.8±0.5ms
               1000000   60.0±0.6ms     26.6±1ms      160±10ms      152±3ms
              ========= ============ ============= ============= =============
```