@pijyoi 2021-08-12
You may want to run it on Qt6.