@j9ac9k 2021-08-12
> You may want to run it on Qt6.

oh you know the issue, I was using `asv run` which doesn't run on your current active branch, it runs on the master branch it would appear...