@j9ac9k 2021-08-12
this PR

```
pyqtgraph-asv ❯ asv run -b arrayToQPath
· Creating environments
· Discovering benchmarks
·· Uninstalling from virtualenv-py3.8-numpy-pyqt6.
·· Building 44e32f00 <qpath_chunks> for virtualenv-py3.8-numpy-pyqt6.
·· Installing 44e32f00 <qpath_chunks> into virtualenv-py3.8-numpy-pyqt6.
· Running 8 total benchmarks (2 commits * 2 environments * 2 benchmarks)
[  0.00%] · For pyqtgraph commit 44e32f00 <qpath_chunks>:
[  0.00%] ·· Building for virtualenv-py3.8-numpy-pyqt6
[  0.00%] ·· Benchmarking virtualenv-py3.8-numpy-pyqt6
[  6.25%] ··· Running (arrayToQPath.TimeSuiteAllFinite.time_test--)..
[ 18.75%] ··· arrayToQPath.TimeSuiteAllFinite.time_test                                              ok
[ 18.75%] ··· ========= ============= ============ ============ ============
              --                               param2
              --------- ----------------------------------------------------
                param1       all         finite       pairs        array
              ========= ============= ============ ============ ============
                10000      147±4μs      149±10μs     921±40μs     922±30μs
                100000   1.33±0.02ms   1.53±0.2ms   8.99±0.3ms   9.11±0.3ms
               1000000    17.2±0.3ms   18.0±0.5ms    130±30ms     103±2ms
              ========= ============= ============ ============ ============

[ 25.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                          ok
[ 25.00%] ··· ========= ============= ============= ============= ============
              --                                param2
              --------- ------------------------------------------------------
                param1       all          finite        pairs        array
              ========= ============= ============= ============= ============
                10000      221±10μs      191±4μs     1.08±0.06ms    977±20μs
                100000   2.75±0.07ms   1.78±0.07ms    9.49±0.2ms   10.1±0.2ms
               1000000    31.3±0.5ms    18.5±0.5ms     144±50ms     103±1ms
              ========= ============= ============= ============= ============

[ 25.00%] ·· Building for virtualenv-py3.8-numpy-pyside6....
[ 25.00%] ·· Benchmarking virtualenv-py3.8-numpy-pyside6
[ 31.25%] ··· Running (arrayToQPath.TimeSuiteAllFinite.time_test--)..
[ 43.75%] ··· arrayToQPath.TimeSuiteAllFinite.time_test                                              ok
[ 43.75%] ··· ========= ============= ============= ============ ============
              --                                param2
              --------- -----------------------------------------------------
                param1       all          finite       pairs        array
              ========= ============= ============= ============ ============
                10000      145±7μs       159±9μs      903±30μs     915±70μs
                100000   1.34±0.02ms   1.43±0.03ms   8.79±0.4ms   8.77±0.2ms
               1000000    17.4±0.9ms    17.8±0.3ms    95.4±4ms    95.0±0.7ms
              ========= ============= ============= ============ ============

[ 50.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                          ok
[ 50.00%] ··· ========= ============ ============= ============ ============
              --                               param2
              --------- ----------------------------------------------------
                param1      all          finite       pairs        array
              ========= ============ ============= ============ ============
                10000     218±5μs       187±7μs      983±30μs     987±30μs
                100000   2.78±0.2ms    1.54±0.2ms   9.48±0.3ms   9.61±0.5ms
               1000000   31.1±0.2ms   17.1±0.02ms    101±1ms     101±0.6ms
              ========= ============ ============= ============ ============
```


master branch

```
[ 50.00%] · For pyqtgraph commit a537bc2b <master>:
[ 50.00%] ·· Building for virtualenv-py3.8-numpy-pyqt6..
[ 50.00%] ·· Benchmarking virtualenv-py3.8-numpy-pyqt6
[ 56.25%] ··· Running (arrayToQPath.TimeSuiteAllFinite.time_test--)..
[ 68.75%] ··· arrayToQPath.TimeSuiteAllFinite.time_test                                              ok
[ 68.75%] ··· ========= ============ ============= ============ ============
              --                               param2
              --------- ----------------------------------------------------
                param1      all          finite       pairs        array
              ========= ============ ============= ============ ============
                10000     167±20μs      159±5μs      939±40μs     914±20μs
                100000   3.03±0.2ms   2.98±0.07ms   10.6±0.5ms   10.2±0.1ms
               1000000    31.2±1ms      33.9±1ms     111±6ms      110±2ms
              ========= ============ ============= ============ ============

[ 75.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                          ok
[ 75.00%] ··· ========= ============ ============ ============= =============
              --                                param2
              --------- -----------------------------------------------------
                param1      all         finite        pairs         array
              ========= ============ ============ ============= =============
                10000     581±30μs     289±20μs    1.54±0.08ms   1.56±0.08ms
                100000   7.46±0.1ms   4.19±0.1ms    16.7±0.3ms    16.6±0.4ms
               1000000    77.7±2ms    42.3±0.9ms     169±6ms       169±2ms
              ========= ============ ============ ============= =============

[ 75.00%] ·· Building for virtualenv-py3.8-numpy-pyside6..
[ 75.00%] ·· Benchmarking virtualenv-py3.8-numpy-pyside6
[ 81.25%] ··· Running (arrayToQPath.TimeSuiteAllFinite.time_test--)..
[ 93.75%] ··· arrayToQPath.TimeSuiteAllFinite.time_test                                              ok
[ 93.75%] ··· ========= ============= ============ ============= ============
              --                                param2
              --------- -----------------------------------------------------
                param1       all         finite        pairs        array
              ========= ============= ============ ============= ============
                10000      174±2μs      185±2μs     1.00±0.02ms    995±20μs
                100000   3.35±0.08ms   2.72±0.7ms    11.8±0.3ms   12.1±0.3ms
               1000000     36.8±1ms     35.3±1ms      118±2ms      114±1ms
              ========= ============= ============ ============= ============

[100.00%] ··· arrayToQPath.TimeSuiteWithNonFinite.time_test                                          ok
[100.00%] ··· ========= ============ ============ ============= ============
              --                               param2
              --------- ----------------------------------------------------
                param1      all         finite        pairs        array
              ========= ============ ============ ============= ============
                10000     644±20μs     334±20μs    1.69±0.08ms   1.83±0.1ms
                100000   8.42±0.3ms   4.64±0.1ms    17.6±0.3ms    18.1±2ms
               1000000   77.2±0.4ms    41.5±1ms      173±2ms      174±4ms
              ========= ============ ============ ============= ============
```